DTS-36327-GSSO security fixes includes: 1.Remove user name paramater from save comment request instead get the user name from  authToken 2.Validating file upload request
